Refactor: 'Samarbeidsplan' funksjoner -> SamarbeidsplanTransactional
